clean up repositories
