+++ app/src/main/java/black/bracken/picsorter/receiver/BootReceiver.kt
@@ -5,24 +5,26 @@ import android.content.Context
 import android.content.Intent
 import black.bracken.picsorter.presentation.observer.ObserverService
 import black.bracken.picsorter.repository.settings.SettingsRepository
+import org.koin.core.KoinComponent
+import org.koin.core.inject
 
 /**
  * @author BlackBracken
  */
-class BootReceiver : BroadcastReceiver() {
+class BootReceiver : BroadcastReceiver(), KoinComponent {
 
-    companion object {
-        const val BOOT_ACTION = "android.intent.action.BOOT_COMPLETED"
-    }
+    private val settingsRepository: SettingsRepository by inject()
 
     override fun onReceive(context: Context, intent: Intent) {
         if (intent.action != BOOT_ACTION) return
 
-        val settingsRepository = SettingsRepository(context)
-
         if (settingsRepository.shouldRunOnBoot) {
             context.startForegroundService(Intent(context, ObserverService::class.java))
         }
+    }
+
+    companion object {
+        const val BOOT_ACTION = "android.intent.action.BOOT_COMPLETED"
     }
 
 }
